Moving json field tests to its own test file + moving django init (DJANGO_SETTINGS_MODULE, django.setup() if needed, ...) to __init__.py file
