+++ tests/models.py
@@ -46,12 +46,3 @@ class SubclassModel(TestModel):
 
 class TestExpressionModel(DirtyFieldsMixin, models.Model):
     counter = models.IntegerField(default=0)
-
-
-try:
-    from jsonfield import JSONField
-
-    class JSONFieldModel(DirtyFieldsMixin, models.Model):
-        json_field = JSONField()
-except ImportError:
-    pass
